+++ node-device/node-plug-state.js
@@ -3,12 +3,12 @@ module.exports = function (RED) {
     RED.nodes.createNode(this, config)
 
     const node = this
-    const device = RED.nodes.getNode(config.device)
-    const plug = RED.nodes.getNode(config.plug)
+    node.device = RED.nodes.getNode(config.device)
+    node.plug = RED.nodes.getNode(config.plug)
 
-    if (device && plug) {
+    if (node.device && node.plug) {
       // Setup a subscriber to the plug's state topic
-      device.subscribe(`plug/${plug.uuid}/state`, node)
+      node.device.subscribe(`plug/${node.plug.uuid}/state`, node)
 
       // Listen to inputs
       node.on('input', function (msg) {
@@ -23,10 +23,10 @@ module.exports = function (RED) {
 
           // Publish the on or off message
           if (typeof onState !== 'undefined') {
-            device.publish(`plug/${plug.uuid}/setstate`, {
+            node.device.publish(`plug/${node.plug.uuid}/setstate`, {
               value: onState ? 'ON' : 'OFF',
               since: new Date().getTime(),
-              nodeId: plug.uuid
+              nodeId: node.plug.uuid
             })
           }
         }