+++ node-device/node-realtime.js
@@ -3,11 +3,11 @@ module.exports = function (RED) {
     RED.nodes.createNode(this, config)
 
     const node = this
-    const device = RED.nodes.getNode(config.device)
+    node.device = RED.nodes.getNode(config.device)
 
-    if (device) {
+    if (node.device) {
       // Setup a subscriber to the device's realtime topic
-      device.subscribe(`realtime`, node)
+      node.device.subscribe(`realtime`, node)
     }
   }